commit
I have updated the UI files for the WeatherApp to accommodate to JavaFX. I do not know where the specific fields for the WeatherApp are, so I could either allow you guys to place them where I have added comments, or you could give me the fields and I could add them to the GUI myself. I leave it up to you, so yeah.